pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "petri-grow"
version = "0.1.0"
description = "An agent orchestration framework to grow your AI's context."
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.9",
    "questionary>=2.0",
    "ollama>=0.4",
    "fastapi>=0.100",
    "sse-starlette>=1.6",
    "uvicorn>=0.23",
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
petri = "petri.cli:app"

[tool.setuptools.packages.find]
include = ["petri*"]
